Backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.5.3
pymongo==4.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
